+++ core/templates/dev/head/pages/exploration_editor/ExplorationEditor.ts
@@ -107,7 +107,7 @@ require(
 require('pages/exploration_editor/ExplorationTitleEditorDirective.ts');
 require('pages/exploration_editor/ParamChangesEditorDirective.ts');
 require('pages/exploration_editor/editor_tab/ExplorationEditorTab.ts');
-require('pages/exploration_editor/editor_tab/ExplorationGraph.ts');
+require('pages/exploration_editor/editor_tab/ExplorationGraphDirective.ts');
 require(
   'pages/exploration_editor/editor_tab/UnresolvedAnswersOverviewDirective.ts');
 require('pages/exploration_editor/feedback_tab/FeedbackTab.ts');
