+++ core/templates/dev/head/pages/exploration_editor/editor_tab/ExplorationGraphDirective.ts
