+++ core/templates/dev/head/pages/exploration_editor/editor_tab/exploration_graph_directive.html
